cleaning up sidebar component
